Add files via upload
MVC is known as an architectural pattern, which embodies three parts Model, View, and Controller, or to be more exact it divides the application into three logical parts: the model part, the view, and the controller. It was used for desktop graphical user interfaces but nowadays is used in designing mobile apps and web apps.